Mitigate /proc/*/environ race condition.
The /proc/*/environ loop was prone to a race condition when processes
disappeared before the shell could redirect the file to tr. This would
make the shell (not tr) emit an error.

This replaces the for loop with a straight "cat /proc/*/environ" which
is then piped to tr. There is still a race, but it's more controlled:
If cat is unable to open a file, it will emit an error and continue to
the next argument. The error is suppressed by redirecting cat's stderr
to /dev/null, which is more obvious that doing the same for the for
loop.

Note: cat should not be prone to hitting ARG_MAX, as it's generally
well above 100'000 even on ancient systems.